+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.16.3 (Unreleased)
+
+- Added quaternion function `to_repr`.
+
+- Added affine function `to_repr`.
+
 ## 0.16.2 (16.2.2026)
 
 - Added affine types: